Fix submit button loading state
On a successful submission, the submit button shows a loader as the form's being submitted, then the user is forwarded to a success page.

Previously the loader stopped showing before the user was forwarded, leading to accidental double submissions, because the user thought the first submission wasn't successful.

Now that doesn't happen.

Fix lint

Fix form submit loader

Fix submit